catpb: use uint64 for privilege bitfield
We are close to running out of bits, since we use a uint32 right now.
This commit buys us more time. If the privilege kinds don't keep
growing too much, then we should be safe for quite a while.

This causes functions that convert protos to json to use double quotes
for the privilege bitfield, since 64-bit integers must be handled as
strings in json.

Release note: None

+++ pkg/sql/privilege/privilege_test.go
@@ -20,7 +20,7 @@ import (
 func TestPrivilegeDecode(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	testCases := []struct {
-		raw              uint32
+		raw              uint64
 		privileges       privilege.List
 		stringer, sorted string
 	}{
